add docstrings for chunky namespace, fix some warnings
