+++ AOEMods.Essence.CLI/Commands.cs
@@ -1,6 +1,5 @@
 ﻿using AOEMods.Essence.Chunky;
 using AOEMods.Essence.Chunky.RGD;
-using AOEMods.Essence.Chunky.RRGeom;
 using AOEMods.Essence.Chunky.RRTex;
 using AOEMods.Essence.SGA;
 using Microsoft.Extensions.FileSystemGlobbing;
@@ -106,7 +105,7 @@ public static class Commands
             if (options.AllMips)
             {
                 string fileName = Path.GetFileName(outputPath);
-                string directoryName = Path.GetDirectoryName(outputPath);
+                string? directoryName = Path.GetDirectoryName(outputPath);
 
                 foreach (var texture in FormatReader.ReadRRTex(rrtexStream, format))
                 {
@@ -118,13 +117,20 @@ public static class Commands
                 try
                 {
                     var texture = FormatReader.ReadRRTexLastMip(rrtexStream, format);
-                    File.WriteAllBytes(outputPath, texture.Value.Data);
+                    if (texture != null)
+                    {
+                        File.WriteAllBytes(outputPath, texture.Value.Data);
+                    }
+                    else if(options.Verbose)
+                    {
+                        Console.WriteLine("No textures could be decoded for {0}", path);
+                    }
                 }
                 catch (InvalidOperationException ex)
                 {
                     if (options.Verbose)
                     {
-                        Console.WriteLine("No textures could be decoded for {0}: {1}", path, ex);
+                        Console.WriteLine("Failed to decode textures for {0}: {1}", path, ex);
                     }
                 }
             }
@@ -206,7 +212,7 @@ public static class Commands
         static void ConvertFile(string path, string outputPath)
         {
             string fileName = Path.GetFileName(outputPath);
-            string directoryName = Path.GetDirectoryName(outputPath);
+            string? directoryName = Path.GetDirectoryName(outputPath);
 
             using var stream = File.Open(path, FileMode.Open, FileAccess.Read);
             int objectIndex = 0;
@@ -218,7 +224,7 @@ public static class Commands
                     FileMode.Create
                 );
 
-                RRGeomUtil.WriteGeometryObject(fileStream, geometryObject);
+                ObjUtil.WriteGeometryObjectAsObj(fileStream, geometryObject);
 
                 objectIndex++;
             }
@@ -238,17 +244,17 @@ public static class Commands
 
     public static int ModelExport(ModelExportOptions options)
     {
-        string materialDirectory = string.IsNullOrEmpty(options.MaterialInputPath) ?
-            (Path.GetDirectoryName(options.InputPath) ?? ".") :
+        string? materialDirectory = string.IsNullOrEmpty(options.MaterialInputPath) ?
+            (Path.GetDirectoryName(options.InputPath)) :
             options.MaterialInputPath;
 
-        string textureDirectory = string.IsNullOrEmpty(options.TextureInputPath) ?
+        string? textureDirectory = string.IsNullOrEmpty(options.TextureInputPath) ?
             materialDirectory :
             options.TextureInputPath;
 
         bool TryFindMaterialPath(string path, out string materialPath)
         {
-            materialPath = Path.Combine(materialDirectory, Path.ChangeExtension(Path.GetRelativePath(options.InputPath, path), ".rrmaterial"));
+            materialPath = Path.Combine(materialDirectory ?? "", Path.ChangeExtension(Path.GetRelativePath(options.InputPath, path), ".rrmaterial"));
             if (File.Exists(materialPath))
             {
                 return true;
@@ -282,14 +288,14 @@ public static class Commands
                     var materials = FormatReader.ReadRRMaterial(materialStream).ToArray();
                     var texturePaths = materials.First(mat => mat.LodTextures.Textures.ContainsKey("albedoTexture") || mat.LodTextures.Textures.ContainsKey("albedoTex")).LodTextures.Textures;
 
-                    var albedoPath = Path.Combine(textureDirectory, texturePaths.ContainsKey("albedoTexture") ? texturePaths["albedoTexture"] : texturePaths["albedoTex"]);
+                    var albedoPath = Path.Combine(textureDirectory ?? "", texturePaths.ContainsKey("albedoTexture") ? texturePaths["albedoTexture"] : texturePaths["albedoTex"]);
                     using var albedoStream = File.OpenRead(albedoPath);
                     var albedoTexture = FormatReader.ReadRRTexLastMip(albedoStream, PngFormat.Instance);
 
                     TextureMip? normalTexture = null;
                     TextureMip? metalTexture = null;
 
-                    var normalPath = Path.Combine(textureDirectory, texturePaths["normalMap"]);
+                    var normalPath = Path.Combine(textureDirectory ?? "", texturePaths["normalMap"]);
                     if (File.Exists(normalPath))
                     {
                         using var normalStream = File.OpenRead(normalPath);
@@ -307,7 +313,7 @@ public static class Commands
             }
 
             string fileName = Path.GetFileName(outputPath);
-            string directoryName = Path.GetDirectoryName(outputPath) ?? ".";
+            string? directoryName = Path.GetDirectoryName(outputPath);
 
             using var stream = File.Open(path, FileMode.Open, FileAccess.Read);
             int objectIndex = 0;
