+++ AOEMods.Essence/Chunky/RGD/RGDList.cs
@@ -1,4 +1,8 @@
-﻿namespace AOEMods.Essence.Chunky;
+﻿namespace AOEMods.Essence.Chunky.RGD;
+
+/// <summary>
+/// List containing key value entries found in the DATA AEGD format of Relic Game Data (RGD) chunks.
+/// </summary>
 public class RGDList : List<KeyValueEntry>
 {
     public RGDList()
