+++ AOEMods.Essence/Chunky/RRGeom/RRGeomDataGeometryC.cs
@@ -1,3 +1,7 @@
 ﻿namespace AOEMods.Essence.Chunky.RRGeom;
 
+/// <summary>
+/// RRGeom DATA GEOC chunk unknown data.
+/// </summary>
+/// <param name="Unknown">Unknown data.</param>
 public record RRGeomDataGeometryC(int[] Unknown);
