+++ AOEMods.Essence/Chunky/RRMaterial/Material.cs
@@ -1,3 +1,7 @@
 ﻿namespace AOEMods.Essence.Chunky.RRMaterial;
 
+/// <summary>
+/// Relic Material (RRMaterial) file material.
+/// </summary>
+/// <param name="LodTextures">Textures contained in the material lod.</param>
 public record Material(RRMaterialPbTextures LodTextures);
