+++ AOEMods.Essence/Chunky/RRMaterial/RRMaterialReader.cs
@@ -4,8 +4,16 @@ using System.Text;
 
 namespace AOEMods.Essence.Chunky.RRMaterial;
 
+/// <summary>
+/// Reads Relic Material (RRMaterial) files.
+/// </summary>
 public class RRMaterialReader : IRRMaterialReader
 {
+    /// <summary>
+    /// Reads Materials from a stream containing an RRMaterial file.
+    /// </summary>
+    /// <param name="stream">Stream containing an RRMaterial file.</param>
+    /// <returns>Materials read from the stream.</returns>
     public static IEnumerable<Material> ReadRRMaterial(Stream stream, string? materialName = null)
     {
         using var reader = new ChunkyFileReader(stream);
